# app/static/swagger.yaml
swagger: '2.0'
info:
  title: Mechanic Shop Final Project
  description: API for managing customers, mechanics, inventory, and service tickets for a mechanic shop.
  version: 1.0.0
host: be-final-modules.onrender.com
basePath: /
schemes:
- https
consumes:
- application/json
produces:
- application/json
securityDefinitions:
  BearerAuth:
    type: apiKey
    name: Authorization
    in: header
    description: 'JWT authorization header. Use: Bearer <token>'
...
paths:
...
  /customers/login:

    post:

      tags:
      - Customers
      summary: Log in as a customer and generate a token

      parameters:
      - in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/LoginPayload'

      responses:
        '200':
          description: Successfully loggin in.
          schema:
            $ref: '#/definitions/LoginResponse'
          examples:
            application/json:
              status: success
              message: Successfully loggin in.
              token: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...

        '400':
          description: "Validation error or invalid email/password. Invalid credentials return: Invalid email or password!"
          schema:
            $ref: '#/definitions/CustomerLoginErrorResponse'
          examples:
            application/json:
              message: "Invalid email or password!"
  /customers/:

    post:

      tags:
      - Customers
      summary: Create a customer account

      parameters:
      - in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/CustomerPayload'

      responses:
        '201':
          description: Customer created.
          schema:
            $ref: '#/definitions/CustomerResponse'
        '400':
          description: Validation error or duplicate email.
          schema:
            $ref: '#/definitions/ErrorResponse'

    get:

      tags:
      - Customers
      summary: Get all customers

      security:
      - BearerAuth: []

      parameters:
      - $ref: '#/parameters/PageParam'
      - $ref: '#/parameters/PerPageParam'

      responses:
        '200':
          description: List of customers.
          schema:
            $ref: '#/definitions/CustomersResponse'
        '401':
          description: Missing or invalid token.
          schema:
            $ref: '#/definitions/ErrorResponse'
        '403':
          description: Insufficient role.
          schema:
            $ref: '#/definitions/ErrorResponse'
  /customers/my-tickets:

    get:

      tags:
      - Customers
      summary: Get service tickets for the logged-in customer

      security:
      - BearerAuth: []

      responses:
        '200':
          description: Customer service tickets.
          schema:
            $ref: '#/definitions/ServiceTicketsResponse'
        '401':
          description: Missing or invalid token.
          schema:
            $ref: '#/definitions/ErrorResponse'
        '403':
          description: Insufficient role.
          schema:
            $ref: '#/definitions/ErrorResponse'
        '404':
          description: Customer not found.
          schema:
            $ref: '#/definitions/ErrorResponse'
  /customers/update_account:

    put:

      tags:
      - Customers
      summary: Update the logged-in customer account

      security:
      - BearerAuth: []

      parameters:
      - in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/CustomerUpdatePayload'

      responses:
        '200':
          description: Updated customer.
          schema:
            $ref: '#/definitions/CustomerResponse'
        '400':
          description: Validation error or duplicate email/phone.
          schema:
            $ref: '#/definitions/ErrorResponse'
        '401':
          description: Missing or invalid token.
          schema:
            $ref: '#/definitions/ErrorResponse'
        '403':
          description: Insufficient role.
          schema:
            $ref: '#/definitions/ErrorResponse'
        '404':
          description: Customer not found.
          schema:
            $ref: '#/definitions/ErrorResponse'
  /customers/delete_account:

    delete:

      tags:
      - Customers
      summary: Delete the logged-in customer account

      security:
      - BearerAuth: []

      responses:
        '200':
          description: Customer deleted.
          schema:
            $ref: '#/definitions/MessageResponse'
        '401':
          description: Missing or invalid token.
          schema:
            $ref: '#/definitions/ErrorResponse'
        '403':
          description: Insufficient role.
          schema:
            $ref: '#/definitions/ErrorResponse'
        '404':
          description: Customer not found.
          schema:
            $ref: '#/definitions/ErrorResponse'
        '409':
          description: Customer has service tickets and cannot be deleted.
          schema:
            $ref: '#/definitions/ErrorResponse'
  /customers/{id}:

    get:

      tags:
      - Customers
      summary: Get one customer by ID

      security:
      - BearerAuth: []

      parameters:
      - $ref: '#/parameters/IdPathParam'

      responses:
        '200':
          description: Customer found.
          schema:
            $ref: '#/definitions/CustomerResponse'
        '401':
          description: Missing or invalid token.
          schema:
            $ref: '#/definitions/ErrorResponse'
        '403':
          description: Insufficient role.
          schema:
            $ref: '#/definitions/ErrorResponse'
        '404':
          description: Customer not found.
          schema:
            $ref: '#/definitions/ErrorResponse'

    put:

      tags:
      - Customers
      summary: Update one customer by ID

      security:
      - BearerAuth: []

      parameters:
      - $ref: '#/parameters/IdPathParam'
      - in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/CustomerUpdatePayload'

      responses:
        '200':
          description: Updated customer.
          schema:
            $ref: '#/definitions/CustomerResponse'
        '400':
          description: Validation error or duplicate email/phone.
          schema:
            $ref: '#/definitions/ErrorResponse'
        '401':
          description: Missing or invalid token.
          schema:
            $ref: '#/definitions/ErrorResponse'
        '403':
          description: Insufficient role.
          schema:
            $ref: '#/definitions/ErrorResponse'
        '404':
          description: Customer not found.
          schema:
            $ref: '#/definitions/ErrorResponse'

    delete:

      tags:
      - Customers
      summary: Delete one customer by ID

      security:
      - BearerAuth: []

      parameters:
      - $ref: '#/parameters/IdPathParam'

      responses:
        '200':
          description: Customer deleted.
          schema:
            $ref: '#/definitions/MessageResponse'
        '401':
          description: Missing or invalid token.
          schema:
            $ref: '#/definitions/ErrorResponse'
        '403':
          description: Insufficient role.
          schema:
            $ref: '#/definitions/ErrorResponse'
        '404':
          description: Customer not found.
          schema:
            $ref: '#/definitions/ErrorResponse'
        '409':
          description: Customer has service tickets and cannot be deleted.
          schema:
            $ref: '#/definitions/ErrorResponse'
...
parameters:
  IdPathParam:
    name: id
    in: path
    required: true
    type: integer
    format: int32
...
  PageParam:
    name: page
    in: query
    required: false
    type: integer
    format: int32
  PerPageParam:
    name: per_page
    in: query
    required: false
    type: integer
    format: int32
definitions:
  LoginPayload:
    type: object
    required:
    - email
    - password
    properties:
      email:
        type: string
        format: email
        example: user@example.com

      password:
        type: string
        format: password
        example: password123
  LoginResponse:
    type: object
    properties:
      status:
        type: string
        example: success

      message:
        type: string
        example: Logged in successfully.

      token:
        type: string
        example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
...
  CustomerLoginErrorResponse:
    type: object
    properties:
      message:
        type: string
        example: Invalid email or password!
...
  CustomerPayload:
    type: object
    required:
    - name
    - email
    - phone
    - password
    properties:
      name:
        type: string
        example: Jane Driver

      email:
        type: string
        format: email
        example: jane@example.com

      phone:
        type: string
        example: 555-0100

      password:
        type: string
        format: password
        example: password123
  CustomerUpdatePayload:
    type: object
    properties:
      name:
        type: string
        example: Jane Driver

      email:
        type: string
        format: email
        example: jane@example.com

      phone:
        type: string
        example: 555-0100

      password:
        type: string
        format: password
        example: newpassword123
  Customer:
    type: object
    properties:
      id:
        type: integer
        example: 1

      name:
        type: string
        example: Jane Driver

      email:
        type: string
        format: email
        example: jane@example.com

      phone:
        type: string
        example: 555-0100

      password:
        type: string
        example: password123
  CustomerResponse:
    allOf:
    - $ref: '#/definitions/Customer'
  CustomersResponse:
    type: array
    items:
      $ref: '#/definitions/Customer'
...
  MechanicSummary:
    type: object
    properties:
      id:
        type: integer
        example: 1

      name:
        type: string
        example: Alex Wrench
...
  InventorySummary:
    type: object
    properties:
      id:
        type: integer
        example: 1

      item_name:
        type: string
        example: Oil Filter

      price:
        type: number
        format: float
        example: 19.99
...
  ServiceTicket:
    type: object
    properties:
      id:
        type: integer
        example: 1

      VIN:
        type: string
        example: 1HGBH41JXMN109186

      service_date:
        type: string
        format: date
        example: '2026-06-17'

      service_desc:
        type: string
        example: Oil change and tire rotation

      customer_id:
        type: integer
        example: 1

      mechanics:
        type: array
        items:
          $ref: '#/definitions/MechanicSummary'

      parts_used:
        type: array
        items:
          $ref: '#/definitions/PartUsed'
...
  ServiceTicketsResponse:
    type: array
    items:
      $ref: '#/definitions/ServiceTicket'
...
  PartUsed:
    type: object
    properties:
      id:
        type: integer
        example: 1

      ticket_id:
        type: integer
        example: 1

      part_id:
        type: integer
        example: 1

      part_quantity:
        type: integer
        example: 2

      part:
        $ref: '#/definitions/InventorySummary'
  MessageResponse:
    type: object
    properties:
      message:
        type: string
        example: Operation completed successfully.
  ErrorResponse:
    type: object
    properties:
      error:
        type: string
        example: Resource not found.

      message:
        type: string
        example: Additional details.
...